make application work in QNX
